refactor: MSW 핸들러 도메인 관심사 분리
