+++ src/mocks/browser.ts
@@ -1,4 +1,4 @@
 import { setupWorker } from 'msw/browser';
-import { handlers } from './handlers';
+import handlers from './handlers';
 
 export const worker = setupWorker(...handlers);
