+++ src/mocks/handlers.ts
@@ -1,18 +1,6 @@
-import { http, HttpResponse } from 'msw';
-import products from './products.json';
-import { API_URL } from '../api/apiClient';
+import { handlers as cartItemHandlers } from './cartItem/cartItem';
+import { handlers as productHandlers } from './product/product';
 
-export const handlers = [
-  http.get(`${API_URL}/products`, ({ request }) => {
-    const url = new URL(request.url);
+const handlers = [...productHandlers, ...cartItemHandlers];
 
-    const page = Number(url.searchParams.get('page') || '1');
-    const limit = page === 0 ? 20 : 4;
-    const start = page === 0 ? 0 : (page - 1) * 4 + 20;
-    const end = start + limit;
-
-    const paginatedProducts = products.slice(start, end);
-
-    return HttpResponse.json({ content: paginatedProducts });
-  }),
-];
+export default handlers;
